src/requirements.txt

numpy
opencv-python
matplotlib
PyYAML
datetime
tqdm
Pillow
pandas